\name{ccTensor-package}
\alias{ccTensor-package}
\alias{ccTensor}
\docType{package}
\title{
\packageTitle{ccTensor}
}
\description{
\packageDescription{ccTensor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ccTensor}
\packageIndices{ccTensor}
}
\author{
\packageAuthor{ccTensor}

Maintainer: \packageMaintainer{ccTensor}
}
\references{
Michael W. Mahoney, et. al., (2009). CUR matrix decompositions for improved data analysis. \emph{PNAS}

Petros Drineas et.al., (2008). Relative-error CUR Matrix Decompositions. \emph{SIAM J. Matrix Anal. Appl.}

Maria F. K. B. et. al. (2019). Multidimensional CX Decomposition of Tensors.
 \emph{WCNPS}
}

\keyword{ package }
\seealso{\code{\link{CX}},\code{\link{CUR}},\code{\link{MultiCX}}}
\examples{
ls("package:ccTensor")
}